requirements.txt
multiqc==1.12
pyarrow==7.0.0
google-cloud-bigquery==2.34.2
pandas==1.3.5
numpy==1.21.5
pytz==2021.3
